Refactor and remove whitespace inconsistencies
