--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Mon Sep 28 16:18:53 BST 2020
-gradle.version=6.5
